gql/parser.go: add comment to checkDependency about change that affects queries
The function checkDependency() checks that variables defined in a query are used, and
that used variables are actually defined. Well this breaks conditional mutations since
we use a query to define variables that are used not in the same query block, but in
a separate mutation block. Nonetheless, the user must be informed of this so a warning
output is generated.